+++ frooky/agent/package.json
@@ -11,8 +11,8 @@
     "build:dev:ios": "node build.js -p ios --type-check none",
     "build:watch:android": "node build.js -p android -t frida -w --type-check none",
     "build:watch:ios": "node build.js -p ios -w -t frida --type-check none",
-    "build:test:android": "frida-compile --type-check none tests/android/runner.ts -o dist/agent-test-android.js",
-    "build:test:ios": "frida-compile --type-check none tests/ios/runner.ts -o dist/agent-test-ios.js",
+    "build:test:android": "frida-compile --type-check none tests/android/agent-runner.ts -o dist/agent-test-android.js",
+    "build:test:ios": "frida-compile --type-check none tests/ios/agent-runner.ts -o dist/agent-test-ios.js",
     "test:android": "npm run build:test:android && node run-tests.js -p android",
     "test:ios": "npm run build:test:ios && node run-tests.js -p ios"
   },
