+++ frooky/agent/tests/ios/agent-runner.ts
@@ -1,10 +1,9 @@
 // tests/ios/test-runner.ts
-import { runTests } from '../test-framework';
+import { runTests } from '../agent-test-framework';
 
 
 // Import all ios test files
 import './test-method-resolver';
-
 
 setTimeout(() => {
     runTests(send);
